backend/requirements.txt
django==2.2.19
djangorestframework==3.13.0
django-filter==2.4.0
djoser==2.0
reportlab==3.6.12